--- ignore.txt
@@ -1,48 +1,0 @@
-.git
-.gitignore
-.gitattributes
-.svn
-__pycache__
-node_modules
-.env
-.idea
-.vscode
-.vagrant
-.vagrantfile
-.terraform
-.terraform.lock.hcl
-.pipenv
-.virtualenv
-.venv
-.dockerignore
-.npmignore
-.gitlab-ci.yml
-.travis.yml
-.jenkinsfile
-.ruby-version
-.rvmrc
-.Gemfile.lock
-.config.ru
-.Capfile
-.yaml
-.log
-.img
-.imgs
-.tmp
-.vendor
-.bin
-.obj
-.build
-.dist
-.out
-.debug
-.release
-.DS_Store
-.Thumbs.db
-.desktop.ini
-.bundle
-.pickle
-.pdf
-.png
-.jpg
-venv
